+++ ogcapi-stable/ogcapi-features-html/src/main/java/de/ii/ldproxy/ogcapi/features/html/app/FeatureCollectionView.java
@@ -54,15 +54,6 @@ public class FeatureCollectionView extends DatasetView {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(FeatureCollectionView.class);
 
-    private class FeatureGeometry {
-        String name;
-        Geometry<?> geometry;
-        FeatureGeometry(String name, Geometry<?> geometry) {
-            this.name = name;
-            this.geometry = geometry;
-        }
-    }
-
     private URI uri;
     public List<NavigationDTO> pagination;
     public List<NavigationDTO> metaPagination;
@@ -81,9 +72,7 @@ public class FeatureCollectionView extends DatasetView {
     public FeaturesHtmlConfiguration.POSITION mapPosition;
     public final MapClient mapClient;
     public final FilterEditor filterEditor;
-    public List<String> geometryProperties;
-    private List<FeatureGeometry> geometries = null;
-    private boolean clampToGround;
+    public final CesiumData cesiumData;
 
     public FeatureCollectionView(OgcApiDataV2 apiData,
         FeatureTypeConfigurationOgcApi collectionData, String template,
@@ -100,8 +89,7 @@ public class FeatureCollectionView extends DatasetView {
         this.schemaOrgFeatures = Objects.nonNull(htmlConfig) && Objects.equals(htmlConfig.getSchemaOrgEnabled(), true);
         this.mapPosition = mapPosition;
         this.uriBuilder = new URICustomizer(uri);
-        this.geometryProperties = geometryProperties;
-        this.clampToGround = true; // TODO make configurable
+        this.cesiumData = new CesiumData(features, geometryProperties);
 
         this.bbox = apiData.getSpatialExtent(collectionData.getId())
             .map(boundingBox -> ImmutableMap.of(
@@ -126,7 +114,6 @@ public class FeatureCollectionView extends DatasetView {
                     .removeZoomLevelConstraints(removeZoomLevelConstraints)
                     .build();
         } else if (mapClientType.equals(MapClient.Type.CESIUM)) {
-            //TODO: Cesium
             this.mapClient = new ImmutableMapClient.Builder()
                 .type(mapClientType)
                 .backgroundUrl(Optional.ofNullable(htmlConfig.getLeafletUrl())
@@ -135,15 +122,6 @@ public class FeatureCollectionView extends DatasetView {
                                        .replace("{y}", "{TileRow}")
                                        .replace("{x}", "{TileCol}")))
                 .attribution(getAttribution())
-                // TODO everything below is ignored / irrelevant
-                .bounds(Optional.ofNullable(bbox))
-                .data(new ImmutableSource.Builder()
-                          .type(TYPE.geojson)
-                          .url(uriBuilder.removeParameters("f").ensureParameter("f", "json").toString())
-                          .build())
-                .popup(Popup.HOVER_ID)
-                .styleUrl(Optional.ofNullable(styleUrl))
-                .removeZoomLevelConstraints(removeZoomLevelConstraints)
                 .build();
         } else {
             LOGGER.error("Configuration error: {} is not a supported map client for the HTML representation of features.", mapClientType);
@@ -171,175 +149,6 @@ public class FeatureCollectionView extends DatasetView {
 
     public boolean isMapRight() {
         return mapPosition == POSITION.RIGHT || (mapPosition == POSITION.AUTO && !features.isEmpty() && features.stream().noneMatch(FeatureHtml::hasObjects));
-    }
-
-    public String getFeatureExtent() {
-        if (Objects.isNull(geometries))
-            geometries = getGeometries(features, geometryProperties);
-        List<Geometry.Coordinate> coordinates = geometries.stream()
-            .map(f -> f.geometry.getCoordinatesFlat())
-            .flatMap(List::stream)
-            .collect(Collectors.toUnmodifiableList());
-        double minLon = getMin(coordinates, 0).orElse(-180.0);
-        double minLat = getMin(coordinates, 1).orElse(-90.0);
-        double maxLon = getMax(coordinates, 0).orElse(180.0);
-        double maxLat = getMax(coordinates, 1).orElse(90.0);
-        return "Cesium.Rectangle.fromDegrees("+minLon+","+minLat+","+maxLon+","+maxLat+");";
-    }
-
-    public List<String> getMapEntities() {
-        ImmutableList.Builder<String> builder = ImmutableList.builder();
-        if (Objects.isNull(geometries))
-            geometries = getGeometries(features, geometryProperties);
-        geometries.stream()
-            .forEach(f -> {
-                final Double minHeight = clampToGround && f.geometry.is3d()
-                    ? getMin(f.geometry.getCoordinatesFlat(), 2).orElse(0.0)
-                    : null;
-                switch (f.geometry.getType()) {
-                    case MultiPoint:
-                        ((Geometry.MultiPoint) f.geometry).getCoordinates()
-                            .stream()
-                            .forEach(point -> addPoint(builder, f.name, point, minHeight));
-                        break;
-                    case Point:
-                        addPoint(builder, f.name, (Geometry.Point) f.geometry, minHeight);
-                        break;
-                    case MultiLineString:
-                        ((Geometry.MultiLineString) f.geometry).getCoordinates()
-                            .stream()
-                            .forEach(line -> addLineString(builder, f.name, line, minHeight));
-                        break;
-                    case LineString:
-                        addLineString(builder, f.name, (Geometry.LineString) f.geometry, minHeight);
-                        break;
-                    case MultiPolygon:
-                        ((Geometry.MultiPolygon) f.geometry).getCoordinates()
-                            .stream()
-                            .forEach(polygon -> addPolygon(builder, f.name, polygon, minHeight));
-                        break;
-                    case Polygon:
-                        addPolygon(builder, f.name, (Geometry.Polygon) f.geometry, minHeight);
-                        break;
-                    default:
-                        throw new IllegalStateException("Unsupported geometry type: " + f.geometry.getType());
-                }
-            });
-        return builder.build();
-    }
-
-    private void addPolygon(ImmutableList.Builder<String> builder, String name, Geometry.Polygon polygon, Double minHeight) {
-        boolean is3d = polygon.is3d();
-        if (polygon.getCoordinates().size()==1) {
-            builder.add("viewer.entities.add({" +
-                            "name:\"" + name + "\"," +
-                            (is3d
-                                ? "polygon:{hierarchy:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(polygon.getCoordinates().get(0).getCoordinates(), minHeight) + "]),perPositionHeight:true,"
-                                : "polygon:{hierarchy:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(polygon.getCoordinates().get(0).getCoordinates()) + "]),perPositionHeight:true,") +
-                            "material:Cesium.Color.BLUE.withAlpha(0.5)," +
-                            "outline:true,outlineColor:Cesium.Color.BLUE" +
-                            "}});");
-        } else {
-            builder.add("viewer.entities.add({" +
-                            "name:\"" + name + "\"," +
-                            (is3d
-                                ? "polygon:{hierarchy:{positions:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(polygon.getCoordinates().get(0).getCoordinates(), minHeight) + "])," +
-                                "holes:[" +
-                                IntStream.range(1, polygon.getCoordinates().size())
-                                    .mapToObj(n -> "{positions:Cesium.Cartesian3.fromDegreesArrayHeights(["+getCoordinatesString(polygon.getCoordinates().get(n).getCoordinates(), minHeight)+"])}")
-                                    .collect(Collectors.joining(",")) +
-                                "]},perPositionHeight:true,"
-                                : "polygon:{hierarchy:{positions:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(polygon.getCoordinates().get(0).getCoordinates()) + "])," +
-                                "holes:[" +
-                                IntStream.range(1, polygon.getCoordinates().size())
-                                    .mapToObj(n -> "{positions:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(polygon.getCoordinates().get(n).getCoordinates()) + "])}")
-                                    .collect(Collectors.joining(",")) +
-                                "]},perPositionHeight:true,") +
-                            "material:Cesium.Color.BLUE.withAlpha(0.5)," +
-                            "outline:true,outlineColor:Cesium.Color.BLUE" +
-                            "}});");
-        }
-    }
-
-    private void addLineString(ImmutableList.Builder<String> builder, String name, Geometry.LineString line, Double minHeight) {
-        builder.add("viewer.entities.add({" +
-                        "name:\"" + name + "\"," +
-                        (line.is3d()
-                            ? "polyline:{positions:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(line.getCoordinates(), minHeight) + "]),perPositionHeight:true,"
-                            : "polyline:{positions:Cesium.Cartesian3.fromDegreesArrayHeights([" + getCoordinatesString(line.getCoordinates()) + "]),perPositionHeight:true,") +
-                        // "outline:true,outlineColor:Cesium.Color.BLUE," +
-                        "width:1," +
-                        "material:Cesium.Color.BLUE" +
-                        "}});");
-
-    }
-
-    private void addPoint(ImmutableList.Builder<String> builder, String name, Geometry.Point point, Double minHeight) {
-        builder.add("viewer.entities.add({" +
-                        "name:\"" + name + "\"," +
-                        "position:Cesium.Cartesian3.fromDegrees(" + getCoordinatesString(point.getCoordinates(), minHeight) + ")," +
-                        "point:{pixelSize:5,color:Cesium.Color.BLUE}" +
-                        "});");
-    }
-
-    private List<FeatureGeometry> getGeometries(List<FeatureHtml> features, List<String> geometryProperties) {
-        return features.stream()
-            .map(feature -> {
-                List<PropertyHtml> geomProperties = ImmutableList.of();
-                for (String geometryProperty : geometryProperties) {
-                    geomProperties = feature.findPropertiesByPath(geometryProperty);
-                    if (!geomProperties.isEmpty())
-                        break;
-                }
-                if (geomProperties.isEmpty()) {
-                    Optional<PropertyHtml> defaultGeom = feature.getGeometry();
-                    if (defaultGeom.isPresent()) {
-                        geomProperties = ImmutableList.of(defaultGeom.get());
-                    }
-                }
-                return geomProperties.stream()
-                    .map(PropertyHtml::parseGeometry)
-                    .map(geom -> new FeatureGeometry(feature.getName(), geom))
-                    .collect(Collectors.toUnmodifiableList());
-            })
-            .flatMap(Collection::stream)
-            .collect(Collectors.toUnmodifiableList());
-    }
-
-    private Optional<Double> getMin(List<Geometry.Coordinate> coordinates, int axis) {
-        return coordinates.stream().map(coord -> coord.get(axis)).min(Comparator.naturalOrder());
-    }
-
-    private Optional<Double> getMax(List<Geometry.Coordinate> coordinates, int axis) {
-        return coordinates.stream().map(coord -> coord.get(axis)).max(Comparator.naturalOrder());
-    }
-
-    private boolean is3d(List<Geometry.Coordinate> coordinates) {
-        return !coordinates.isEmpty() && coordinates.get(0).size()==3;
-    }
-
-    private String getCoordinatesString(List<Geometry.Coordinate> coordinates) {
-        if (is3d(coordinates))
-            return coordinates.stream()
-                .flatMap(List::stream)
-                .map(String::valueOf)
-                .collect(Collectors.joining(","));
-
-        return coordinates.stream()
-            .map(coord -> ImmutableList.of(coord.get(0), coord.get(1), 0.0))
-            .flatMap(List::stream)
-            .map(String::valueOf)
-            .collect(Collectors.joining(","));
-    }
-
-    private String getCoordinatesString(List<Geometry.Coordinate> coordinates, Double deltaHeight) {
-        if (Objects.isNull(deltaHeight))
-            return getCoordinatesString(coordinates);
-        return coordinates.stream()
-            .map(coord -> Geometry.Coordinate.of(coord.get(0), coord.get(1), coord.get(2) - deltaHeight))
-            .flatMap(List::stream)
-            .map(String::valueOf)
-            .collect(Collectors.joining(","));
     }
 
     @Override
