+++ ogcapi-stable/ogcapi-html/src/main/javascript/src/apps/cesium/index.jsx
@@ -1,0 +1,6 @@
+/* eslint-disable no-undef, no-underscore-dangle */
+import Cesium from "../../components/Cesium";
+
+if (globalThis._map && globalThis._cesium && globalThis._map.container) {
+  Cesium({ ...globalThis._map, ...globalThis._cesium });
+}
