+++ ogcapi-stable/ogcapi-html/src/main/javascript/src/components/Cesium/style.css
@@ -1,0 +1,8 @@
+.cesium-viewer-bottom {
+  left: 5px !important;
+  bottom: 0 !important;
+}
+
+.cesium-viewer-bottom a:nth-child(2) {
+  display: none !important;
+}
