+++ ogcapi-stable/ogcapi-html/src/main/resources/templates/htmlStart.mustache
@@ -21,12 +21,6 @@
 
 {{> style-common}}
 
-<!-- TODO temporary hack -->
-{{#mapClient}}{{#isCesium}}
-<script src="https://cesium.com/downloads/cesiumjs/releases/1.84/Build/Cesium/Cesium.js"></script>
-<link href="https://cesium.com/downloads/cesiumjs/releases/1.84/Build/Cesium/Widgets/widgets.css" rel="stylesheet">
-{{/isCesium}}{{/mapClient}}
-
 {{> schemaOrg}}
 
 </head>
