+++ ogcapi-stable/ogcapi-html/src/main/resources/templates/mapClient.mustache
@@ -77,36 +77,9 @@
     {{> app-openlayers}}
   {{/isOpenLayers}}
   {{#isCesium}}
-    {{> app-maplibre}}
+    {{> data-cesium}}
+    {{> app-cesium}}
   {{/isCesium}}
-
-<!-- TODO temporary hack -->
-{{#isCesium}}
-<script>
-Cesium.Ion.defaultAccessToken = null;
-Cesium.Camera.DEFAULT_VIEW_RECTANGLE = {{{featureExtent}}};
-Cesium.Camera.DEFAULT_VIEW_FACTOR = 0;
-var viewer = new Cesium.Viewer("map", {
-    {{#backgroundUrl}}
-    imageryProvider : new Cesium.WebMapTileServiceImageryProvider({ url : '{{{.}}}' }),
-    {{/backgroundUrl}}
-    animation : false,
-    baseLayerPicker : false,
-    homeButton : true,
-    geocoder : false,
-    timeline : false,
-    scene3DOnly : true,
-    fullscreenButton : false,
-    navigationInstructionsInitiallyVisible : false
-});
-{{#attribution}}
-var credit = new Cesium.Credit('{{{.}}}');
-viewer.scene.frameState.creditDisplay.addDefaultCredit(credit);
-{{/attribution}}
-{{#mapEntities}}{{{.}}}{{/mapEntities}}
-viewer.zoomTo(viewer.entities);
-</script>
-{{/isCesium}}
 
 {{/mapClient}}
